Create README.md
Hello! This is Fawwaz Adil, a freshie in BSCS at The FAST, NUCES-Karachi campus. This directory is designed for my first semester's Programming Fundamentals course in which I would be working, and saving all of my files/code;
My passion for system management and problem-solving is deep-rooted since the beginning of my O-Levels, where I had the opportunity to participate in a Hackathon, and engineer a proto-type to solve the issue of climate change. I hope one day I can make this happen in real life, and be a form of benefit for humanity some way or the other.